pyproject.toml
[tool.poetry]
name = "icon-event-publisher"
version = "0.1.0"
description = ""
authors = ["Brian Li <brian@brianli.com>"]

[tool.poetry.dependencies]
python = "3.10.2"
iconsdk = "^2.0.0"
rich = "^11.1.0"
pydantic = "^1.9.0"
requests = "^2.27.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
